move endpoint /me from auth to profile; fix imports
